ref: Rename file to react router v3
