+++ packages/react/test/reactrouterv3.test.tsx
@@ -2,7 +2,7 @@ import { render } from '@testing-library/react';
 import * as React from 'react';
 import { createMemoryHistory, createRoutes, IndexRoute, match, Route, Router } from 'react-router-3';
 
-import { Match, reactRouterV3Instrumentation, Route as RouteType } from '../src/reactrouter';
+import { Match, reactRouterV3Instrumentation, Route as RouteType } from '../src/reactrouterv3';
 
 // Have to manually set types because we are using package-alias
 declare module 'react-router-3' {
